requirements.txt
Flask
Pillow
requests
torch
transformers
